+++ docs/release_notes.rst
@@ -5,6 +5,7 @@ v0.1 - Unreleased
 -----------------
 
 - Documentation update and tutorial.
+- Allow rechunk to accept a Dask array.
 
 
 v0.0.1 - 2020-07-15